requirements.txt
streamlit
opencv-python-headless
torch
torchvision
Pillow
ultralytics